Use different image for removal tests
When removing an image in tests, remove alpinelinux/base instead of
busybox:buildroot-2013.08.1. This is a similarly small image.

This doesn't change the efficacy of our tests at all, but it does let them
run on CircleCI, where there is some weird interaction between the rootfs
of busybox and the modifications that the CircleCI folks have made to the
ZFS driver.